requirements.txt
django==1.3.1
django-annoying
django-piston
django-registration
